[spring-server-with-federation-as-a-subgraph] refactor : 패키지 정리
