+++ 놀이터(예제 코드 작성)/graphql-federation/spring-server-with-federation-as-a-subgraph/src/main/kotlin/com/banjjoknim/subgraphservice/domain/book/api/BookDtos.kt
@@ -1,6 +1,6 @@
 package com.banjjoknim.subgraphservice.domain.book.api
 
-import com.banjjoknim.subgraphservice.domain.book.datasource.Book
+import com.banjjoknim.subgraphservice.domain.book.model.Book
 
 data class BookResponse(
     val number: Int,
