+++ 놀이터(예제 코드 작성)/graphql-federation/spring-server-with-federation-as-a-subgraph/src/main/kotlin/com/banjjoknim/subgraphservice/domain/book/infrastructure/datasource/BookDataSource.kt
@@ -1,4 +1,6 @@
-package com.banjjoknim.subgraphservice.domain.book.datasource
+package com.banjjoknim.subgraphservice.domain.book.infrastructure.datasource
+
+import com.banjjoknim.subgraphservice.domain.book.model.Book
 
 object BookDataSource {
