+++ 놀이터(예제 코드 작성)/graphql-federation/spring-server-with-federation-as-a-subgraph/src/main/kotlin/com/banjjoknim/subgraphservice/domain/book/infrastructure/messaging/kafka/KafkaConfiguration.kt
@@ -1,4 +1,4 @@
-package com.banjjoknim.subgraphservice.config.messaging.kafka
+package com.banjjoknim.subgraphservice.domain.book.infrastructure.messaging.kafka
 
 import org.apache.kafka.clients.admin.AdminClientConfig
 import org.apache.kafka.clients.admin.NewTopic
