+++ 놀이터(예제 코드 작성)/graphql-federation/spring-server-with-federation-as-a-subgraph/src/main/kotlin/com/banjjoknim/subgraphservice/domain/book/infrastructure/messaging/redis/RedisChannel.kt
@@ -1,5 +1,7 @@
-package com.banjjoknim.subgraphservice.domain.book.infrastructure
+package com.banjjoknim.subgraphservice.domain.book.infrastructure.messaging.redis
 
+import com.banjjoknim.subgraphservice.domain.book.infrastructure.messaging.MessageChannel
+import com.banjjoknim.subgraphservice.domain.book.infrastructure.messaging.PubSubMessage
 import org.springframework.context.annotation.Profile
 import org.springframework.data.redis.core.ReactiveStringRedisTemplate
 import org.springframework.stereotype.Component
