+++ 놀이터(예제 코드 작성)/graphql-federation/spring-server-with-federation-as-a-subgraph/src/main/kotlin/com/banjjoknim/subgraphservice/domain/book/model/Book.kt
@@ -1,4 +1,4 @@
-package com.banjjoknim.subgraphservice.domain.book.datasource
+package com.banjjoknim.subgraphservice.domain.book.model
 
 data class Book(
     val number: Int,
